Fix larastan level 5 errors: Fix Money object null coalescing in models

+++ app/Filament/Clusters/Accounting/Resources/VendorBills/Pages/EditVendorBill.php
@@ -241,6 +241,9 @@ class EditVendorBill extends EditRecord
 
             DeleteAction::make()
                 ->action(function (Model $record) {
+                    if (!$record instanceof \App\Models\VendorBill) {
+                        throw new \Exception('Invalid record type');
+                    }
                     app(VendorBillService::class)->delete($record);
                     $this->redirect(VendorBillResource::getUrl('index'));
                 }),
